+++ tests/unit/utilities/test_attachments_api.py
@@ -12,7 +12,7 @@ class MockResponse:
         self.mode = mode
 
     def json(self):
-        if self.mode == "get_run_ids":
+        if self.mode == "get_run_ids_from_response":
             return {
                 "value": [
                     {"id": 1},
@@ -77,78 +77,50 @@ def test_print_response_info_does_not_print_when_response_has_no_title(mock_prin
 
 
 @mock.patch("builtins.print")
-def test_parse_parameters_returns_valid_params(mock_print):
-    args = ["unused_arg", "test-release-id", "test-auth-token"]
-
-    params = parse_parameters(args)
-
-    assert_that(params, equal_to(("test-release-id", "test-auth-token")), "Incorrect parameters returned")
-    expected_message = "The release ID is: test-release-id"
-    mock_print.assert_called_with(expected_message)
-
-
-@mock.patch("builtins.print")
-def test_parse_parameters_handles_invalid_number_of_arguments(mock_print):
-    args = ["test1", "test2"]
-
-    params = parse_parameters(args)
-
-    assert_that(params, equal_to(None), "No parameters should be returned when arguments are invalid")
-    expected_message = "##vso[task.logissue type=error]Unable to parse required parameters - " \
-                       "ensure they are passed correctly. Expected release ID and auth token."
-    mock_print.assert_called_with(expected_message)
-
-
-@mock.patch("builtins.print")
-@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 200, "get_run_ids"))
-def test_get_run_ids_performs_the_correct_request(mock_get, _mock_print):
-    get_run_ids(100, "test-url", "test-token")
+@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 200, "get_run_ids_from_response"))
+def test_get_run_ids_by_release_performs_the_correct_request(mock_get, _mock_print):
+    get_run_ids_by_release(100, "test-url", "test-token")
 
     check_mocked_functions_called(mock_get)
     request_args = mock_get.call_args
 
     assert_that(request_args[0][0], equal_to("test-url"), "request_url incorrect")
     assert_that(request_args[1]["params"]["releaseIds"], equal_to(100), "releaseIds incorrect")
-    assert_that(request_args[1]["params"]["api-version"], equal_to("5.0"), "api-version for GET incorrect")
+    assert_that(request_args[1]["params"]["api-version"], equal_to("6.0"), "api-version for GET incorrect")
     assert_that(request_args[1]["auth"][1], equal_to("test-token"), "Auth token incorrect")
 
 
 @mock.patch("builtins.print")
-@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 200, "get_run_ids"))
-def test_get_run_ids_uses_the_correct_min_and_max_dates_in_its_response(mock_get, _mock_print):
-    get_run_ids(100, "test-url", "test-token")
+@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 200, "get_run_ids_from_response"))
+def test_get_run_ids_by_build_performs_the_correct_request(mock_get, _mock_print):
+    get_run_ids_by_build(100, "test-url", "test-token")
 
     check_mocked_functions_called(mock_get)
     request_args = mock_get.call_args
 
-    expected_date_time = datetime.datetime.now()
-    actual_date_time = request_args[1]["params"]["maxLastUpdatedDate"]
-    time_difference = (actual_date_time - expected_date_time).total_seconds()
-    assert_that(time_difference, less_than(2), "maxLastUpdatedDate incorrect")
-
-    expected_date_time = get_date_altered_by_days(-1, datetime.datetime.now())
-    actual_date_time = request_args[1]["params"]["minLastUpdatedDate"]
-    time_difference = (actual_date_time - expected_date_time).total_seconds()
-    assert_that(time_difference, less_than(2), "minLastUpdatedDate incorrect")
+    assert_that(request_args[0][0], equal_to("test-url"), "request_url incorrect")
+    assert_that(request_args[1]["params"]["buildIds"], equal_to(100), "buildIds incorrect")
+    assert_that(request_args[1]["params"]["api-version"], equal_to("6.0"), "api-version for GET incorrect")
+    assert_that(request_args[1]["auth"][1], equal_to("test-token"), "Auth token incorrect")
 
 
 @mock.patch("builtins.print")
-@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 200, "get_run_ids"))
-def test_get_run_ids_returns_ids_when_request_succeeds(mock_get, _mock_print):
-    run_ids = get_run_ids(100, "test-url", "test-token")
+@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 200, "get_run_ids_from_response"))
+def test_get_run_ids_by_release_returns_ids_when_request_succeeds(mock_get, _mock_print):
+    run_ids = get_run_ids_by_release(100, "test-url", "test-token")
 
     check_mocked_functions_called(mock_get)
     assert_that(run_ids, equal_to([1, 2, 3]), "Incorrect run IDs returned")
 
 
-@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 400, "get_run_ids"))
+@mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 400, "get_run_ids_from_response"))
 @mock.patch("builtins.print")
-def test_get_run_ids_returns_no_ids_when_request_fails(mock_print, mock_get):
-    run_ids = get_run_ids(100, "test-url", "test-token")
+def test_get_run_ids_by_release_returns_no_ids_when_request_fails(mock_print, mock_get):
+    run_ids = get_run_ids_by_release(100, "test-url", "test-token")
 
     check_mocked_functions_called(mock_get, mock_print)
     assert_that(run_ids, equal_to([]), "Run IDs should not be returned when the request failed")
-    mock_print.assert_called_with("##vso[task.logissue type=error]Could not get run IDs for release 100")
+    mock_print.assert_called_with("##vso[task.logissue type=error]Could not get run IDs for release/build 100")
 
 
 @mock.patch("requests.get", side_effect=lambda *args, **kwargs: MockResponse("", 400, "get_failed_tests"))
@@ -181,7 +153,7 @@ def test_get_failed_tests_performs_the_correct_request(mock_get, _mock_print):
     request_args = mock_get.call_args
 
     assert_that(request_args[0][0], equal_to("test-url/100/results"), "request_url incorrect")
-    assert_that(request_args[1]["params"]["api-version"], equal_to("5.0"), "api-version for GET incorrect")
+    assert_that(request_args[1]["params"]["api-version"], equal_to("6.0"), "api-version for GET incorrect")
     assert_that(request_args[1]["auth"][1], equal_to("test-token"), "Auth token incorrect")
 
 
@@ -201,108 +173,133 @@ def test_attach_files_fails_when_passed_no_parameters(_mock_print):
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: None)
-def test_attach_files_fails_when_there_are_no_run_ids(mock_get_run_ids, _mock_print):
-    result = attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: None)
+def test_attach_files_release_fails_when_there_are_no_run_ids(mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
-    check_mocked_functions_called(mock_get_run_ids)
+    check_mocked_functions_called(mock_get_run_ids_from_response)
     assert_that(result, equal_to(1), "Result should be a failure when there are no run IDs")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
-@mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: None)
-def test_attach_files_fails_when_there_are_no_failed_tests(mock_get_failed_tests, mock_get_run_ids, _mock_print):
-    result = attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: None)
+def test_attach_files_build_fails_when_there_are_no_run_ids(mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_build("test-org", "test-project", "screenshots", "100", "test-token")
 
-    check_mocked_functions_called(mock_get_failed_tests, mock_get_run_ids)
+    check_mocked_functions_called(mock_get_run_ids_from_response)
+    assert_that(result, equal_to(1), "Result should be a failure when there are no run IDs")
+
+
+@mock.patch("builtins.print")
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: None)
+def test_attach_files_fails_when_there_are_no_failed_tests(mock_get_failed_tests, mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
+
+    check_mocked_functions_called(mock_get_failed_tests, mock_get_run_ids_from_response)
     assert_that(result, equal_to(1), "Result should be a failure when there are no failed tests")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
 @mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: None)
 @mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"],
                                                                                                 [11, 101, "test2"]])
-def test_attach_files_fails_when_there_are_no_files(mock_get_failed_tests, mock_listdir, mock_get_run_ids, _mock_print):
-    result = attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+def test_attach_files_fails_when_there_are_no_files(mock_get_failed_tests, mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
-    check_mocked_functions_called(mock_listdir, mock_get_failed_tests, mock_get_run_ids)
+    check_mocked_functions_called(mock_listdir, mock_get_failed_tests, mock_get_run_ids_from_response)
     assert_that(result, equal_to(1), "Result should be a failure when there are no files")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
 @mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: ["test1", "test2"])
 @mock.patch("uitestcore.utilities.attachments_api.get_file_base64", side_effect=lambda *args: None)
 @mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"],
                                                                                                 [11, 101, "test2"]])
 def test_attach_files_fails_when_there_base64_conversion_fails(mock_get_failed_tests, mock_get_file_base64,
-                                                               mock_listdir, mock_get_run_ids, _mock_print):
-    result = attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+                                                               mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
-    check_mocked_functions_called(mock_get_failed_tests, mock_get_file_base64, mock_listdir, mock_get_run_ids)
+    check_mocked_functions_called(mock_get_failed_tests, mock_get_file_base64, mock_listdir, mock_get_run_ids_from_response)
     assert_that(result, equal_to(1), "Result should be a failure when base64 conversion fails")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
 @mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: ["file1"])
 @mock.patch("uitestcore.utilities.attachments_api.get_file_base64", side_effect=lambda *args: None)
 @mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"]])
 def test_attach_files_gets_the_base64_of_the_correct_file(mock_get_failed_tests, mock_get_file_base64,
-                                                          mock_listdir, mock_get_run_ids, _mock_print):
-    attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+                                                          mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
-    check_mocked_functions_called(mock_get_failed_tests, mock_listdir, mock_get_run_ids)
+    check_mocked_functions_called(mock_get_failed_tests, mock_listdir, mock_get_run_ids_from_response)
     mock_get_file_base64.assert_called_with("screenshots/test1/file1")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
 @mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: ["test1", "test2"])
 @mock.patch("uitestcore.utilities.attachments_api.get_file_base64", side_effect=lambda *args: b"test-base64-string")
 @mock.patch("requests.post", side_effect=lambda *args, **kwargs: MockResponse("", 400))
 @mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"],
                                                                                                 [11, 101, "test2"]])
 def test_attach_files_fails_when_the_request_fails(mock_get_failed_tests, mock_post, mock_get_file_base64,
-                                                   mock_listdir, mock_get_run_ids, _mock_print):
-    result = attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+                                                   mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
     check_mocked_functions_called(mock_get_failed_tests, mock_post, mock_get_file_base64,
-                                  mock_listdir, mock_get_run_ids)
+                                  mock_listdir, mock_get_run_ids_from_response)
     assert_that(result, equal_to(1), "Result should be a failure when the request fails")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
 @mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: ["test1", "test2"])
 @mock.patch("uitestcore.utilities.attachments_api.get_file_base64", side_effect=lambda *args: b"test-base64-string")
 @mock.patch("requests.post", side_effect=lambda *args, **kwargs: MockResponse("", 200))
 @mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"],
                                                                                                 [11, 101, "test2"]])
-def test_attach_files_succeeds(mock_get_failed_tests, mock_post, mock_get_file_base64,
-                               mock_listdir, mock_get_run_ids, _mock_print):
-    result = attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+def test_attach_files_release_succeeds(mock_get_failed_tests, mock_post, mock_get_file_base64,
+                               mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
     check_mocked_functions_called(mock_get_failed_tests, mock_post, mock_get_file_base64,
-                                  mock_listdir, mock_get_run_ids)
+                                  mock_listdir, mock_get_run_ids_from_response)
     assert_that(result, equal_to(0), "Result should be a success when everything works")
 
 
 @mock.patch("builtins.print")
-@mock.patch("uitestcore.utilities.attachments_api.get_run_ids", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
+@mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: ["test1", "test2"])
+@mock.patch("uitestcore.utilities.attachments_api.get_file_base64", side_effect=lambda *args: b"test-base64-string")
+@mock.patch("requests.post", side_effect=lambda *args, **kwargs: MockResponse("", 200))
+@mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"],
+                                                                                                [11, 101, "test2"]])
+def test_attach_files_build_succeeds(mock_get_failed_tests, mock_post, mock_get_file_base64,
+                               mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    result = attach_files_build("test-org", "test-project", "screenshots", "100", "test-token")
+
+    check_mocked_functions_called(mock_get_failed_tests, mock_post, mock_get_file_base64,
+                                  mock_listdir, mock_get_run_ids_from_response)
+    assert_that(result, equal_to(0), "Result should be a success when everything works")
+
+
+@mock.patch("builtins.print")
+@mock.patch("uitestcore.utilities.attachments_api.get_run_ids_from_response", side_effect=lambda *args: [10, 11])
 @mock.patch("uitestcore.utilities.attachments_api.listdir", side_effect=lambda *args: ["test1", "test2"])
 @mock.patch("uitestcore.utilities.attachments_api.get_file_base64", side_effect=lambda *args: b"test-base64-string")
 @mock.patch("requests.post", side_effect=lambda *args, **kwargs: MockResponse("", 200))
 @mock.patch("uitestcore.utilities.attachments_api.get_failed_tests", side_effect=lambda *args: [[10, 100, "test1"],
                                                                                                 [11, 101, "test2"]])
 def test_attach_files_performs_the_correct_request(mock_get_failed_tests, mock_post, mock_get_file_base64,
-                                                   mock_listdir, mock_get_run_ids, _mock_print):
-    attach_files("test-org", "test-project", "screenshots", [None, "100", "test-token"])
+                                                   mock_listdir, mock_get_run_ids_from_response, _mock_print):
+    attach_files_release("test-org", "test-project", "screenshots", "100", "test-token")
 
     check_mocked_functions_called(mock_get_failed_tests, mock_post, mock_get_file_base64,
-                                  mock_listdir, mock_get_run_ids)
+                                  mock_listdir, mock_get_run_ids_from_response)
     request_args = mock_post.call_args
 
     assert_that(request_args[0][0], equal_to("https://dev.azure.com/test-org/test-project/_apis/"